re-enable '/' cmds and update index.js file to re-depreciation warning
Prettier inserted a fair number of semi-colons and trailing commas
I can remove them if requested.

I left a commented out section in the messages JSX as I am not sure
that my styling of the emotes will be to everyone's taste. Currently
name and message text are displayed in a single line in italics. I
thought it redundant to include the author name in bold in the case of
an emote.

For system messages I did not update the styling so they currently do
not indent or have a 'cabal' logo of any sort.