Implement cabal-client granual events
The single `update` event from cabal-client was creating a loop of events and slowing the Desktop client to a stop. This implements cabal-client's newer individual events which updates the UI only when it should. This should hold us for a while until I can finish the `cabal-ui` implementation.

I also fixed a bug that would post to the wrong channel after switching between cabals.