electron: set enableRemoteModule to true (default changed in v9)
"In Electron 9, using the remote module without explicitly enabling it via the enableRemoteModule WebPreferences option began emitting a warning. In Electron 10, the remote module is now disabled by default. To use the remote module, enableRemoteModule: true must be specified in WebPreferences:"

https://github.com/electron/electron/blob/main/docs/breaking-changes.md#default-changed-enableremotemodule-defaults-to-false

if migrating to larger versions, in particular for electron-store@2 see the following links

https://stackoverflow.com/questions/55850616/uncaught-referenceerror-module-is-not-defined-on-delete-module-exports
https://github.com/electron/remote/issues/79
https://github.com/electron/remote/blob/main/docs/migration-2.md